Reduce percieved layout shift from loading body font
The final body font loaded is 'Source Sans Pro', a sans-serif font. However, while the font is being downloaded, the site renders serif fonts. On a slow or busy network this can lead to a dramatic shift in font style of body content as it transitions from a serif font to a sans-serif font.

This change attempts to reduce the impact of that shift by providing fallback fonts that are also sans-serif, so the body content transitions from one sans-serif font to another sans-serif font