Improve initial load of FontAwesome icons (#231)
* Improve initial load of FontAwesome icons

Use a gatsby plugin to improve the inital load of font awesome icons. This plugin inlines the FontAwesome CSS in the head of the HTML pages so that when the SVGs are rendered they are properly styled and sized on initial render.

* Add gatsby clean command

* Avoid slight layout shift in header by adding width size to logo

This change ensures the properly sized box is reserved by the browser's layout engine before the logo image downloads avoiding a slight content shift as the image downloads

* Reduce percieved layout shift from loading body font

The final body font loaded is 'Source Sans Pro', a sans-serif font. However, while the font is being downloaded, the site renders serif fonts. On a slow or busy network this can lead to a dramatic shift in font style of body content as it transitions from a serif font to a sans-serif font.

This change attempts to reduce the impact of that shift by providing fallback fonts that are also sans-serif, so the body content transitions from one sans-serif font to another sans-serif font

* Move gatsby plugin to dependencies