Why I am late to this meeting (version 1)
Aparently, by following the advice in this documentation, I got off when the US switched to daylight savings time. The meeting is in fact, not 1900 UTC, but 1800 UTC (during US DST).

This version of the fix adjusts the meeting back by one hour UTC.

This alternative PR fixes the anchor time zone to US Pacific time. Please choose one :)