Add recent OpenUI discussion results
The largest change is the result of [this discussion](https://github.com/openui/open-ui/issues/581) about the differences between dialog and popup=manual. There's a new section attempting to describe those differences. This PR also includes recent changes to the `show` event, and an update on the interaction between top layer element types.

This PR also removes an obsolete section of the `<selectmenu>`
page that talks about replaceable shadow DOM.