fix: migrate component description to definition
As mentioned in https://github.com/openui/open-ui/pull/760#pullrequestreview-1530855140, component data that used the term "description" was migrated to use "definition" as defined in the schema. This will cause the definition to show up on the research pages.

Also, a few misspellings of "decription" were cleaned up with this change.