Clean up source data (#760)
This cleans up some of the json data in the sources by:
* renaming description to definition on the component level
* removing category as it's unused
* fixing typos, e.g. decription to description and switching invalid
  property names to their possibly intended valid counterpart
* pointing datasets to correct schema
* marking the image property on the concepts level as optional (about 
  576 images are missing currently)
* removing duplicate properties
* fixing indentation

Some descriptions that were previously not displayed should now be 
visible. Also, some data that was not visible on the research pages due
to typos should now be displayed.

This also reduces the number of JSON schema warnings to 1 (a missing URL 
for the chromium file upload where no useful URL can be found as it's
the native component viewed in a chromium browser).